Move paas metrics cloudwatch service into own pkg
